requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
html5lib==1.1
idna==2.10
numpy==1.19.2
Pillow==7.2.0
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
webencodings==0.5.1
